move some static method to AggregationUtil
